// src/app/models/item.ts
export interface ItemData {
    item_id: string;
    buy_price_max: number;
    buy_price_max_date: string;
    buy_price_min: number;
    buy_price_min_date: string;
    city: string;
    quality: number;
    sell_price_max: number;
    sell_price_max_date: string;
    sell_price_min: number;
    sell_price_min_date: string;
}